+++ src/UHC/Util/CHR.hs
@@ -1,0 +1,11 @@
+
+-- | Wrapper module around CHR stuff
+
+module UHC.Util.CHR
+  ( module UHC.Util.CHR.Base
+  , module UHC.Util.CHR.Key
+  )
+  where
+
+import UHC.Util.CHR.Base
+import UHC.Util.CHR.Key